# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: CIVORA API
  description: |
    CIVORA API - open digital infrastructure for public-interest service delivery.

    Version: 0.2.0-service-delivery

    ## Public-Interest Service Delivery Lifecycle

    Full end-to-end workflow for public service delivery:
    Request → Person registered → Case created (NEW) → OPEN →
    Eligibility assessed → Evidence collected → IN_REVIEW →
    Assessment completed → DECISION_PENDING → Decision made →
    APPROVED/REJECTED → IN_PROGRESS → Assistance provided →
    FOLLOW_UP → CLOSED

    State machine:
    NEW → OPEN → IN_REVIEW → ASSESSMENT → DECISION_PENDING →
    APPROVED → IN_PROGRESS → FOLLOW_UP → CLOSED
    REJECTED → CLOSED
  version: "0.2.0"
  contact:
    name: CIVORA
    url: https://github.com/alrazihi/civora
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080
    description: Local development server. Override with CIVORA_API_URL in production.
security: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT bearer token. The token must include `organization_id` claim.
        The API validates that the token's organization matches the `{orgId}`
        path parameter on all tenant-scoped endpoints.
...
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: INVALID_INPUT
            message:
              type: string
              example: "title is required"
...
    Evidence:
      type: object
      required: [id, organization_id, service_request_id, type, description, storage_reference, uploaded_by, created_at]
      properties:
        id:
          type: string
          format: uuid
          pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
        organization_id:
          type: string
          format: uuid
          pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
        service_request_id:
          type: string
          format: uuid
          pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
        type:
          type: string
          enum: [IDENTITY_DOCUMENT, PROOF_OF_RESIDENCE, REFERRAL, SUPPORTING_DOCUMENT, PHOTOGRAPH, STAFF_NOTE]
        description:
          type: string
        storage_reference:
          type: string
        uploaded_by:
          type: string
          format: uuid
          pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
        created_at:
          type: string
          format: date-time
    CreateEvidenceRequest:
      type: object
      required: [service_request_id, type, description, storage_reference]
      properties:
        service_request_id:
          type: string
          format: uuid
          pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
        type:
          type: string
          enum: [IDENTITY_DOCUMENT, PROOF_OF_RESIDENCE, REFERRAL, SUPPORTING_DOCUMENT, PHOTOGRAPH, STAFF_NOTE]
        description:
          type: string
        storage_reference:
          type: string
...
    FollowUp:
      type: object
      required: [id, organization_id, service_request_id, scheduled_date, outcome, notes, performed_by, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
          pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
        organization_id:
          type: string
          format: uuid
          pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
        service_request_id:
          type: string
          format: uuid
          pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
        scheduled_date:
          type: string
          format: date
        completed_date:
          type: string
          format: date
          nullable: true
        outcome:
          type: string
        notes:
          type: string
        performed_by:
          type: string
          format: uuid
          pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateFollowUpRequest:
      type: object
      required: [service_request_id, scheduled_date, outcome, notes]
      properties:
        service_request_id:
          type: string
          format: uuid
          pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
        scheduled_date:
          type: string
          format: date
          description: "ISO 8601 date (YYYY-MM-DD)"
        outcome:
          type: string
        notes:
          type: string
    CompleteFollowUpRequest:
      type: object
      required: [completed_date]
      properties:
        completed_date:
          type: string
          format: date
          description: "ISO 8601 date (YYYY-MM-DD)"
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
paths:
...
  /organizations/{orgId}/evidence:
    post:
      operationId: createOrgevidence
      tags: [Evidence]
      summary: Add evidence
      description: Adds an evidence item for a service request. Requires authentication.
      security:
        - BearerAuth: []
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
            format: uuid
            pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEvidenceRequest'
      responses:
        '201':
          description: Evidence created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Evidence'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /organizations/{orgId}/evidence/{evidenceId}:
    get:
      operationId: getEvidenceById
      tags: [Evidence]
      summary: Get evidence
      description: Retrieves a specific evidence item. Requires authentication.
      security:
        - BearerAuth: []
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
            format: uuid
            pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
        - name: evidenceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
            pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
      responses:
        '200':
          description: Evidence details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Evidence'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /organizations/{orgId}/evidence/by-service-request/{serviceRequestId}:
    get:
      operationId: listEvidenceByServiceRequestId
      tags: [Evidence]
      summary: List evidence by service request
      description: Lists evidence items for a specific service request. Requires authentication.
      security:
        - BearerAuth: []
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
            format: uuid
            pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
        - name: serviceRequestId
          in: path
          required: true
          schema:
            type: string
            format: uuid
            pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 20
      responses:
        '200':
          description: List of evidence items with pagination
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Evidence'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /organizations/{orgId}/follow-ups:
    post:
      operationId: createOrgfollowups
      tags: [FollowUps]
      summary: Create a follow-up
      description: Creates a new follow-up item for a service request. Requires authentication.
      security:
        - BearerAuth: []
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
            format: uuid
            pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFollowUpRequest'
      responses:
        '201':
          description: Follow-up created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FollowUp'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /organizations/{orgId}/follow-ups/{followUpId}:
    get:
      operationId: getFollowUpById
      tags: [FollowUps]
      summary: Get a follow-up
      description: Retrieves a specific follow-up item. Requires authentication.
      security:
        - BearerAuth: []
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
            format: uuid
            pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
        - name: followUpId
          in: path
          required: true
          schema:
            type: string
            format: uuid
            pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
      responses:
        '200':
          description: Follow-up details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FollowUp'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /organizations/{orgId}/follow-ups/{followUpId}/complete:
    patch:
      operationId: completeFollowUp
      tags: [FollowUps]
      summary: Complete a follow-up
      description: Marks a follow-up as completed. Requires admin or staff role.
      security:
        - BearerAuth: []
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
            format: uuid
            pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
        - name: followUpId
          in: path
          required: true
          schema:
            type: string
            format: uuid
            pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteFollowUpRequest'
      responses:
        '200':
          description: Completed follow-up
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FollowUp'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Forbidden - admin or staff role required
          content:
            application/json:
              schema:
                 $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/NotFound'
  /organizations/{orgId}/follow-ups/by-service-request/{serviceRequestId}:
    get:
      operationId: listFollowUpsByServiceRequestId
      tags: [FollowUps]
      summary: List follow-ups by service request
      description: Lists follow-up items for a specific service request. Requires authentication.
      security:
        - BearerAuth: []
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
            format: uuid
            pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
        - name: serviceRequestId
          in: path
          required: true
          schema:
            type: string
            format: uuid
            pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 20
      responses:
        '200':
          description: List of follow-up items with pagination
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FollowUp'
        '401':
          $ref: '#/components/responses/Unauthorized'